+++ Assets/OxGFrame/AssetLoader/ThirdParty/YooAsset/Runtime/PackageSystem/PackageBundle.cs
@@ -75,12 +75,13 @@ namespace YooAsset
 				if (IsRawFile)
 				{
 					string cacheRoot = PersistentTools.GetCachedRawFileFolderPath(PackageName);
-					_cachedDataFilePath = $"{cacheRoot}/{folderName}/{CacheGUID}/{YooAssetSettings.CacheBundleDataFileName}{_fileExtension}";
+					_cachedDataFilePath = PathUtility.Combine(cacheRoot, folderName, CacheGUID, YooAssetSettings.CacheBundleDataFileName);
+					_cachedDataFilePath += _fileExtension;
 				}
 				else
 				{
 					string cacheRoot = PersistentTools.GetCachedBundleFileFolderPath(PackageName);
-					_cachedDataFilePath = $"{cacheRoot}/{folderName}/{CacheGUID}/{YooAssetSettings.CacheBundleDataFileName}";
+					_cachedDataFilePath = PathUtility.Combine(cacheRoot, folderName, CacheGUID, YooAssetSettings.CacheBundleDataFileName);
 				}
 				return _cachedDataFilePath;
 			}
@@ -101,12 +102,12 @@ namespace YooAsset
 				if (IsRawFile)
 				{
 					string cacheRoot = PersistentTools.GetCachedRawFileFolderPath(PackageName);
-					_cachedInfoFilePath = $"{cacheRoot}/{folderName}/{CacheGUID}/{YooAssetSettings.CacheBundleInfoFileName}";
+					_cachedInfoFilePath = PathUtility.Combine(cacheRoot, folderName, CacheGUID, YooAssetSettings.CacheBundleInfoFileName);
 				}
 				else
 				{
 					string cacheRoot = PersistentTools.GetCachedBundleFileFolderPath(PackageName);
-					_cachedInfoFilePath = $"{cacheRoot}/{folderName}/{CacheGUID}/{YooAssetSettings.CacheBundleInfoFileName}";
+					_cachedInfoFilePath = PathUtility.Combine(cacheRoot, folderName, CacheGUID, YooAssetSettings.CacheBundleInfoFileName);
 				}
 				return _cachedInfoFilePath;
 			}